requirements.txt
google-cloud-storage==3.1.0